requirements.txt
discord.py
python-dotenv
Flask
pytest
pytest-asyncio
pytest-mock
pytest-cov
Jinja2
pytz
aiohttp
requests
supabase